+++ files/Providers/AuthServiceProvider.php
@@ -1,7 +1,0 @@
-<?php namespace App\Providers;
-
-use Eyewill\TucleCore\Providers\AuthServiceProvider as ServiceProvider;
-
-class AuthServiceProvider extends ServiceProvider
-{
-}