+++ files/Providers/EventServiceProvider.php
@@ -1,7 +1,0 @@
-<?php namespace App\Providers;
-
-use Eyewill\TucleCore\Providers\EventServiceProvider as ServiceProvider;
-
-class EventServiceProvider extends ServiceProvider
-{
-}