+++ files/Providers/RouteServiceProvider.php
@@ -1,7 +1,0 @@
-<?php namespace App\Providers;
-
-use Eyewill\TucleCore\Providers\RouteServiceProvider as ServiceProvider;
-
-class RouteServiceProvider extends ServiceProvider
-{
-}