+++ src/Contracts/Eloquent/SortableInterface.php
@@ -1,0 +1,6 @@
+<?php namespace Eyewill\TucleCore\Contracts\Eloquent;
+
+interface SortableInterface
+{
+  public function sortOrder($order = -1);
+}